backend/requirements.txt
annotated-types>=0.7.0
anyio>=4.0.0
attrs>=23.0.0
bcrypt>=4.0.0
certifi>=2024.0.0
cffi>=1.16.0
charset-normalizer>=3.3.0
click>=8.1.0
colorama>=0.4.6
coloredlogs>=15.0.0
cryptography>=41.0.0
dnspython>=2.6.0
ecdsa>=0.19.0
email-validator>=2.1.0
fastapi>=0.110.0
flatbuffers>=24.0.0
google-ai-generativelanguage>=0.6.0
google-api-core>=2.18.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.8.0
googleapis-common-protos>=1.63.0
greenlet>=3.0.0
grpcio>=1.62.0
grpcio-status>=1.62.0
h11>=0.14.0
httpcore>=1.0.0
httplib2>=0.22.0
httpx>=0.27.0
humanfriendly>=10.0
idna>=3.6
ImageIO>=2.34.0
jsonschema>=4.21.0
jsonschema-specifications>=2023.12.0
lazy_loader>=0.4
llvmlite>=0.42.0
mpmath>=1.3.0
networkx>=3.3
numba>=0.59.0
numpy>=1.26.0
onnxruntime>=1.17.0
packaging>=24.0
passlib>=1.7.4
pillow>=10.3.0
platformdirs>=4.2.0
pooch>=1.8.0
proto-plus>=1.23.0
protobuf>=5.26.0
pyasn1>=0.6.0
pyasn1_modules>=0.4.0
pycparser>=2.22
pydantic>=2.7.0
pydantic-settings>=2.2.0
pydantic_core>=2.18.0
PyMatting>=1.1.12
pyparsing>=3.1.0
pyreadline3>=3.4
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.9
referencing>=0.35.0
requests>=2.31.0
rpds-py>=0.18.0
rsa>=4.9
scikit-image>=0.23.0
scipy>=1.13.0
six>=1.16.0
SQLAlchemy>=2.0.0
starlette>=0.37.0
sympy>=1.12
tifffile>=2024.1.0
tqdm>=4.66.0
typing-inspection>=0.4.0
typing_extensions>=4.10.0
uritemplate>=4.1.1
urllib3>=2.2.0
uvicorn>=0.29.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
